move const data to code_803E724
