--- data/data_80F59C8_0.s
@@ -1,62 +1,0 @@
-#include "constants/status.h"
-.section .rodata
-#.incbin "baserom.gba", 0xF59C4, 0x4
-
-.string "pksdir0\0"
-.global gUnknown_80F6490
-gUnknown_80F6490: @ 80F6490
-@ replacing .incbin "baserom.gba", 0x000f6490, 0x24
-.byte 0x00, 0x00, 0x00, 0x0
-.byte 0x0f, 0x00, 0x00, 0x00
-.byte 0xff, 0x00, 0x00, 0x00
-.byte 0xff, 0x0f, 0x00, 0x00
-.byte 0xff, 0xff, 0x00, 0x00
-.byte 0xff, 0xff, 0x0f, 0x00
-.byte 0xff, 0xff, 0xff, 0x00
-.byte 0xff, 0xff, 0xff, 0x0f
-.byte 0xff, 0xff, 0xff, 0xff
-
-.global gUnknown_80F64B4
-gUnknown_80F64B4: @ 80F64B4
-@ replacing .incbin "baserom.gba", 0x000f64b4, 0x24
-.byte 0x5a, 0x00, 0x00, 0x00
-.byte 0x50, 0x00, 0x00, 0x00
-.byte 0x46, 0x00, 0x00, 0x00
-.byte 0x3c, 0x00, 0x00, 0x00
-.byte 0x32, 0x00, 0x00, 0x00
-.byte 0x28, 0x00, 0x00, 0x00
-.byte 0x1e, 0x00, 0x00, 0x00
-.byte 0x14, 0x00, 0x00, 0x00
-.byte 0x0a, 0x00, 0x00, 0x00
-
-.global gUnknown_80F64D8
-gUnknown_80F64D8: @ 80F64D8
-@ replacing .incbin "baserom.gba", 0x000f64d8, 0x24
-.byte 0x91, 0xf2, 0x90, 0xf2, 0x8f, 0xf2, 0x8e, 0xf2, 0x8d, 0xf2, 0x8c, 0xf2, 0x8b, 0xf2, 0x8a, 0xf2, 0x89, 0xf2, 0xc8, 0xf2, 0xc7, 0xf2, 0xc6, 0xf2, 0xc5, 0xf2, 0xc4, 0xf2, 0xc3, 0xf2, 0xc2, 0xf2
-.byte 0xc1, 0xf2, 0xc0, 0xf2
-
-.global gUnknown_80F64FC
-gUnknown_80F64FC: @ 80F64FC
-@ replacing .incbin "baserom.gba", 0x000f64fc, 0x24
-.byte 0x84, 0x03, 0x00, 0x00, 0x20, 0x03, 0x00, 0x00, 0xbc, 0x02, 0x00, 0x00, 0x58, 0x02, 0x00, 0x00, 0xf4, 0x01, 0x00, 0x00, 0x90, 0x01, 0x00, 0x00, 0x2c, 0x01, 0x00, 0x00, 0xc8, 0x00, 0x00, 0x00
-.byte 0x64, 0x00, 0x00, 0x00
-
-.global gUnknown_80F6520
-gUnknown_80F6520: @ 80F6520
-@ replacing .incbin "baserom.gba", 0x000f6520, 0x24
-.byte 0x5a, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x46, 0x00, 0x00, 0x00, 0x3c, 0x00, 0x00, 0x00, 0x32, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x1e, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00, 0x00
-.byte 0x0a, 0x00, 0x00, 0x00
-
-.global gUnknown_80F6544
-gUnknown_80F6544: @ 80F6544
-@ replacing .incbin "baserom.gba", 0x000f6544, 0x24
-.byte 0x91, 0xf2, 0x90, 0xf2, 0x8f, 0xf2, 0x8e, 0xf2, 0x8d, 0xf2, 0x8c, 0xf2, 0x8b, 0xf2, 0x8a, 0xf2, 0x89, 0xf2, 0xc8, 0xf2, 0xc7, 0xf2, 0xc6, 0xf2, 0xc5, 0xf2, 0xc4, 0xf2, 0xc3, 0xf2, 0xc2, 0xf2
-.byte 0xc1, 0xf2, 0xc0, 0xf2
-
-.global gUnknown_80F6568
-gUnknown_80F6568: @ 80F6568
-@ replacing .incbin "baserom.gba", 0x000f6568, 0x88
-.byte 0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x01, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x01, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0x01, 0x00, 0x00, 0x00, 0xfe, 0xff, 0xff, 0xff
-.byte 0x01, 0x00, 0x00, 0x00, 0xfe, 0xff, 0xff, 0xff, 0x02, 0x00, 0x00, 0x00, 0xfe, 0xff, 0xff, 0xff, 0x02, 0x00, 0x00, 0x00, 0xfe, 0xff, 0xff, 0xff, 0x02, 0x00, 0x00, 0x00, 0xfd, 0xff, 0xff, 0xff
-.byte 0x03, 0x00, 0x00, 0x00, 0xfd, 0xff, 0xff, 0xff, 0x03, 0x00, 0x00, 0x00, 0xfd, 0xff, 0xff, 0xff, 0x04, 0x00, 0x00, 0x00, 0xfc, 0xff, 0xff, 0xff, 0x04, 0x00, 0x00, 0x00, 0xfc, 0xff, 0xff, 0xff
-.byte 0x04, 0x00, 0x00, 0x00, 0xfb, 0xff, 0xff, 0xff, 0x04, 0x00, 0x00, 0x00, 0xfb, 0xff, 0xff, 0xff, 0x04, 0x00, 0x00, 0x00, 0xfb, 0xff, 0xff, 0xff, 0x04, 0x00, 0x00, 0x00, 0xfb, 0xff, 0xff, 0xff
